+++ src/com/koxudaxi/ruff/RuffConfigService.kt
@@ -5,12 +5,18 @@ import com.intellij.openapi.components.State
 import com.intellij.openapi.components.Storage
 import com.intellij.openapi.project.Project
 import com.intellij.util.xmlb.XmlSerializerUtil
+import org.jetbrains.annotations.SystemDependent
 
 @State(name = "RuffConfigService", storages = [Storage("ruff.xml")])
 class RuffConfigService : PersistentStateComponent<RuffConfigService> {
     var runRuffOnSave: Boolean = false
     var runRuffOnReformatCode: Boolean = true
     var showRuleCode: Boolean = false
+    var globalRuffExecutablePath: @SystemDependent String? = null
+    var alwaysUseGlobalRuff: Boolean = false
+    var projectRuffExecutablePath: @SystemDependent String? = null
+
+
     override fun getState(): RuffConfigService {
         return this
     }
@@ -18,7 +24,13 @@ class RuffConfigService : PersistentStateComponent<RuffConfigService> {
     override fun loadState(config: RuffConfigService) {
         XmlSerializerUtil.copyBean(config, this)
     }
-
+    val ruffExecutablePath: @SystemDependent String?
+        get() {
+        return when {
+            alwaysUseGlobalRuff -> globalRuffExecutablePath
+            else -> projectRuffExecutablePath ?: globalRuffExecutablePath
+        }
+    }
     companion object {
         fun getInstance(project: Project): RuffConfigService {
             return project.getService(RuffConfigService::class.java)